+++ HISTORY.md
@@ -72,6 +72,7 @@
 
 ### 2022-12-02
 * gui: added configuration option for "blocksZoom"
+* gui: added user-event for dropping a costume icon in the wardrobe
 
 ### 2022-12-01
 * blocks: don't show "result pic" option in context menu with "click-to-run" disabled